requirements.txt
pandas
numpy
seaborn
scikit-learn
dvc
mlflow
xgboost
ipykernel
python-dotenv
imblearn
ensure
python-box
flask
dagshub
dvc_s3
-e .